@@ -26,12 +26,14 @@
   "devDependencies": {
     "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
+    "@eslint/js": "^9.1.1",
     "@vue/cli-plugin-babel": "^5.0.4",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "^5.0.4",
     "@vue/compiler-sfc": "^3.2.31",
-    "eslint": "^8.12.0",
-    "eslint-plugin-vue": "^9.0.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.25.0",
+    "globals": "^15.1.0",
     "sass": "^1.42.1",
     "sass-loader": "^14.1.0"
   },
