@@ -1,4 +1,4 @@
-module.exports = {
+export const plugins = {
   presets: [
     '@vue/cli-plugin-babel/preset'
   ]
