@@ -1,4 +1,5 @@
-const { defineConfig } = require('@vue/cli-service')
+import { defineConfig } from '@vue/cli-service'
+
 module.exports = defineConfig({
   transpileDependencies: true
 })
